Prototype MySQL writer is now working.
Had to add it to the list of known writers and fix a bunch of small bugs
in my original code. Currently, there is a warning on shutdown about a
thread not exiting. Since I haven't seen this before, I assume it's
because of something I did (probably the MySQL driver which isn't
explicitly closed). Will have to investigate further later.